success

+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: edcastro <edcastro@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/18 12:23:06 by edcastro          #+#    #+#             */
-/*   Updated: 2023/05/23 16:40:16 by edcastro         ###   ########.fr       */
+/*   Updated: 2023/06/19 18:05:41 by edcastro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
 	size_t	i;
